Split cdx_writer.py into mulitple modules
CDX-Writer now internally consists of multiple modules in cdx_writer
package.
cdx_writer.py no longer exists in the source tree, and cdx_writer script
is generated by pip.
This is a change necessary to allow 3rd party code to import extension
point classes.
Besides change to deployment requeirement and script name, it's
compatible with previous versions.
