+++ cdx_writer/__main__.py
@@ -1,0 +1,4 @@
+import sys
+from cdx_writer.command import main
+
+exit(main(sys.argv[1:]))
